oauthutil: Give context to check for second
The SecondAttemt method gets the existing token to check if a second
attempt is needed.
I didn't really think about the old implementation, it could only work
on the first login attempt, since it modified the Scope directly inside
the provider struct. This mustn't happen, it affect all requests except
the first one. So yeah, it worked, but only for the first user...

Also: The second request allows modification of the LoginURL, the
redirect will pass a second bool to the LoginURL method. The method can
then decide if it wants to modify its request.